third_party/blink/web_tests/fast/dom/Document/document-close-iframe-load-expected.txt

ALERT: PASS (just to make expected.txt non-empty)
